+++ parser/src/error.rs
@@ -8,6 +8,7 @@ mod duplicate_service_uuid;
 mod duplicate_struct_field;
 mod duplicate_struct_field_id;
 mod empty_enum;
+mod expected_type_found_const;
 mod expected_type_found_service;
 mod import_not_found;
 mod invalid_const_value;
@@ -38,6 +39,7 @@ pub use duplicate_service_uuid::DuplicateServiceUuid;
 pub use duplicate_struct_field::DuplicateStructField;
 pub use duplicate_struct_field_id::DuplicateStructFieldId;
 pub use empty_enum::EmptyEnum;
+pub use expected_type_found_const::ExpectedTypeFoundConst;
 pub use expected_type_found_service::ExpectedTypeFoundService;
 pub use import_not_found::ImportNotFound;
 pub use invalid_const_value::InvalidConstValue;
@@ -68,6 +70,7 @@ pub enum Error {
     DuplicateStructField(DuplicateStructField),
     DuplicateStructFieldId(DuplicateStructFieldId),
     EmptyEnum(EmptyEnum),
+    ExpectedTypeFoundConst(ExpectedTypeFoundConst),
     ExpectedTypeFoundService(ExpectedTypeFoundService),
     ImportNotFound(ImportNotFound),
     InvalidConstValue(InvalidConstValue),
@@ -104,6 +107,7 @@ impl Diagnostic for Error {
             Self::DuplicateStructField(e) => e.schema_name(),
             Self::DuplicateStructFieldId(e) => e.schema_name(),
             Self::EmptyEnum(e) => e.schema_name(),
+            Self::ExpectedTypeFoundConst(e) => e.schema_name(),
             Self::ExpectedTypeFoundService(e) => e.schema_name(),
             Self::ImportNotFound(e) => e.schema_name(),
             Self::InvalidConstValue(e) => e.schema_name(),
@@ -136,6 +140,7 @@ impl Diagnostic for Error {
             Self::DuplicateStructField(e) => e.format(parsed),
             Self::DuplicateStructFieldId(e) => e.format(parsed),
             Self::EmptyEnum(e) => e.format(parsed),
+            Self::ExpectedTypeFoundConst(e) => e.format(parsed),
             Self::ExpectedTypeFoundService(e) => e.format(parsed),
             Self::ImportNotFound(e) => e.format(parsed),
             Self::InvalidConstValue(e) => e.format(parsed),