+++ parser/src/error/expected_type_found_const.rs
@@ -1,0 +1,103 @@
+use super::Error;
+use crate::ast::{NamedRef, NamedRefKind};
+use crate::diag::{Diagnostic, DiagnosticKind, Formatted, Formatter};
+use crate::validate::Validate;
+use crate::{util, Parsed};
+
+#[derive(Debug)]
+pub struct ExpectedTypeFoundConst {
+    schema_name: String,
+    named_ref: NamedRef,
+    candidate: Option<String>,
+}
+
+impl ExpectedTypeFoundConst {
+    pub(crate) fn validate(named_ref: &NamedRef, validate: &mut Validate) {
+        let (schema, ident) = match named_ref.kind() {
+            NamedRefKind::Intern(ident) => (validate.get_current_schema(), ident),
+
+            NamedRefKind::Extern(schema, ident) => {
+                let Some(schema) = validate.get_schema(schema.value()) else {
+                    return;
+                };
+
+                (schema, ident)
+            }
+        };
+
+        let mut found = false;
+        for def in schema.definitions() {
+            if def.name().value() == ident.value() {
+                if def.as_const().is_some() {
+                    found = true;
+                } else {
+                    return;
+                }
+            }
+        }
+
+        if found {
+            let candidate =
+                util::did_you_mean_type(schema, ident.value(), true).map(ToOwned::to_owned);
+
+            validate.add_error(Self {
+                schema_name: validate.schema_name().to_owned(),
+                named_ref: named_ref.clone(),
+                candidate,
+            });
+        }
+    }
+
+    pub fn named_ref(&self) -> &NamedRef {
+        &self.named_ref
+    }
+}
+
+impl Diagnostic for ExpectedTypeFoundConst {
+    fn kind(&self) -> DiagnosticKind {
+        DiagnosticKind::Error
+    }
+
+    fn schema_name(&self) -> &str {
+        &self.schema_name
+    }
+
+    fn format<'a>(&'a self, parsed: &'a Parsed) -> Formatted<'a> {
+        let mut fmt = Formatter::new(
+            self,
+            format!(
+                "expected type; found constant `{}`",
+                self.named_ref.ident().value()
+            ),
+        );
+
+        if let Some(schema) = parsed.get_schema(&self.schema_name) {
+            fmt.main_block(
+                schema,
+                self.named_ref.span().from,
+                self.named_ref.span(),
+                "type expected here",
+            );
+        }
+
+        if let Some(ref candidate) = self.candidate {
+            match self.named_ref.schema() {
+                Some(schema) => {
+                    fmt.help(format!("did you mean `{}::{candidate}`?", schema.value()));
+                }
+
+                None => {
+                    fmt.help(format!("did you mean `{candidate}`?"));
+                }
+            }
+        }
+
+        fmt.format()
+    }
+}
+
+impl From<ExpectedTypeFoundConst> for Error {
+    fn from(e: ExpectedTypeFoundConst) -> Self {
+        Self::ExpectedTypeFoundConst(e)
+    }
+}