+++ parser/src/test/ui_tests.rs
@@ -1,4 +1,5 @@
 ui_test!(error_empty_enum);
+ui_test!(expected_type_found_const);
 ui_test!(expected_type_found_service);
 ui_test!(grapheme_width);
 ui_test!(invalid_event_id);